dom/push/test/test_push_manager_worker.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1184574: Expose PushManager to workers.

Any copyright is dedicated to the Public Domain.
http://creativecommons.org/licenses/publicdomain/

-->
<head>
  <title>Test for Bug 1184574</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8">
</head>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1184574">Mozilla Bug 1184574</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="text/javascript">

  var registration;

  function start() {
    return navigator.serviceWorker.register("worker.js" + "?" + (Math.random()), {scope: "."})
    .then(swr => { registration = swr; return swr; });
  }

  function unregisterSW() {
    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    }, function(e) {
      dump("Unregistering the SW failed with " + e + "\n");
    });
  }

  function setupPushNotification(swr) {
    return swr.pushManager.subscribe().then(
      pushSubscription => {
        ok(true, "successful registered for push notification");
        return pushSubscription;
      }, error => {
        ok(false, "could not register for push notification");
      });
  }

  function getNewEndpointFromWorker(pushSubscription) {
    return new Promise((resolve, reject) => {
      var channel = new MessageChannel();
      channel.port1.onmessage = e => {
        (e.data.error ? reject : resolve)(e.data);
      };
      registration.active.postMessage({
        endpoint: pushSubscription.endpoint,
      }, [channel.port2]);
    }).then(data => {
      return registration.pushManager.getSubscription().then(
        pushSubscription => {
          is(data.endpoint, pushSubscription.endpoint,
             "Wrong push endpoint in parent");
          return pushSubscription;
      });
    });
  }

  function unregisterPushNotification(pushSubscription) {
    return pushSubscription.unsubscribe().then(
      result => {
      ok(result, "unsubscribe() on existing subscription should return true.");
      return pushSubscription;
    }, error => {
      ok(false, "unsubscribe() should never fail.");
    });
  }

  function runTest() {
    start()
    .then(setupPushNotification)
    .then(getNewEndpointFromWorker)
    .then(unregisterPushNotification)
    .then(unregisterSW)
    .catch(function(e) {
      ok(false, "Some test failed with error " + e);
    }).then(SimpleTest.finish);
  }

  SpecialPowers.pushPrefEnv({"set": [
    ["dom.push.enabled", true],
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
    ]}, runTest);
  SpecialPowers.addPermission('push', true, document);
  SimpleTest.waitForExplicitFinish();
</script>
</body>
</html>
